Abstracted BlockProperty
